+++ strutil/strutil_test.go
@@ -64,7 +64,7 @@ func (s *StrUtilSuite) TestSubstring(c *C) {
 	c.Assert(Substring("test1234TEST", 4, 8), Equals, "1234")
 	c.Assert(Substring("test1234TEST", 4, 4), Equals, "")
 	c.Assert(Substring("test1234TEST", 8, 100), Equals, "TEST")
-	c.Assert(Substring("test1234TEST", 6, -10), Equals, "test12")
+	c.Assert(Substring("test1234TEST", 2, -6), Equals, "st12")
 	c.Assert(Substring("简单的消息", -1, 2), Equals, "简单")
 	c.Assert(Substring("Пример", -1, 2), Equals, "Пр")
 	c.Assert(Substring("test1234TEST", 12, 99), Equals, "")