+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **`[log]`** Print messages with `AUX` level as `INFO` for JSON format
 - **`[strutil]`** Added support of negative end value
+- **`[knf/value]`** Code refactoring
 
 ### [13.23.1](https://kaos.sh/ek/13.23.1)